Move to own module, simplify logic, document
Also add remaining geometries
