+++ geo/src/lib.rs
@@ -136,6 +136,7 @@
 //! - **[`HaversineIntermediate`](HaversineIntermediate)**:
 //! - **[`Proj`](Proj)**: Project geometries with the `proj` crate (requires the `use-proj` feature)
 //! - **[`ChaikinSmoothing`](ChaikinSmoothing)**: Smoothen `LineString`, `Polygon`, `MultiLineString` and `MultiPolygon` using Chaikins algorithm.
+//! - **[`Densify`](Densify)**: Densify linear geometry components by interpolating points
 //!
 //! # Features
 //!
